refactor: deleteOne()
